pyproject.toml
[project]
name = "tiktok-bot-template"
version = "0.1.0"
description = "Template proyek untuk bot TikTok menggunakan Python dan Playwright."
authors = [{ name = "Jules", email = "jules@example.com" }]
requires-python = ">=3.11"
dependencies = [
    "playwright",
    "python-dotenv",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [
  "."
]